orangecarrier_cli_bot/requirements.txt

playwright
python-dotenv
requests
